add: 詳細画面、編集画面でお気に入り状態を変更できるよう実装 #34
