+++ front/src/components/elements/IdeaMemoCard/IdeaMemoCard.module.scss
@@ -18,79 +18,6 @@
   }
 }
 
-.cardHeader {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  width: 100%;
-  height: 80px;
-  padding: 16px;
-
-  @media screen and (min-width: 768px) {
-    padding: 20px;
-  }
-}
-
-.heartLike {
-  height: 32px;
-  width: auto;
-  cursor: pointer;
-  transition: transform 0.25s ease-in-out;
-  transform: scale(1);
-}
-
-.heartUnlike {
-  height: 32px;
-  width: auto;
-  cursor: pointer;
-  transition: transform 0.25s ease-in-out;
-}
-
-.heartBeat {
-  animation: heartBeat 0.25s ease-in-out;
-}
-
-@keyframes heartBeat {
-  50% {
-    transform: scale(1.2);
-  }
-}
-
-.cardBody {
-  width: 100%;
-  padding: 16px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 16px;
-
-  @media screen and (min-width: 768px) {
-    padding: 20px;
-  }
-}
-
-.section {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 8px;
-  width: 100%;
-}
-
-.sectionContentContainer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  width: 100%;
-}
-
-.sectionContent {
-  width: 98%;
-  text-align: start;
-}
-
 .cardFooter {
   width: 100%;
   height: 80px;
