+++ front/src/components/elements/IdeaMemoCardContent/IdeaMemoCardContent.module.scss
@@ -1,0 +1,34 @@
+.cardBody {
+  width: 100%;
+  padding: 16px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 16px;
+
+  @media screen and (min-width: 768px) {
+    padding: 20px;
+  }
+}
+
+.section {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  gap: 8px;
+  width: 100%;
+}
+
+.sectionContentContainer {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  width: 100%;
+}
+
+.sectionContent {
+  width: 98%;
+  text-align: start;
+}
