+++ front/src/components/elements/IdeaMemoCardFooter/IdeaMemoCardFooter.module.scss
