+++ front/src/components/elements/IdeaMemoCardHeader/IdeaMemoCardHeader.module.scss
@@ -1,0 +1,37 @@
+.cardHeader {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  width: 100%;
+  height: 80px;
+  padding: 16px;
+
+  @media screen and (min-width: 768px) {
+    padding: 20px;
+  }
+}
+
+.heartLike {
+  height: 32px;
+  width: auto;
+  cursor: pointer;
+  transition: transform 0.25s ease-in-out;
+  transform: scale(1);
+}
+
+.heartUnlike {
+  height: 32px;
+  width: auto;
+  cursor: pointer;
+  transition: transform 0.25s ease-in-out;
+}
+
+.heartBeat {
+  animation: heartBeat 0.25s ease-in-out;
+}
+
+@keyframes heartBeat {
+  50% {
+    transform: scale(1.2);
+  }
+}
