+++ front/src/components/elements/Search/Search.module.scss
@@ -1,7 +1,12 @@
 .container {
   box-sizing: border-box;
   position: relative;
-  width: 80%;
+  width: 100%;
+  margin-right: 8px;
+
+  @media screen and (min-width: 768px) {
+    margin-right: 16px;
+  }
 }
 
 .label {
