--- front/src/features/idea-memos-uuid/components/IdeaMemoCardHeader/IdeaMemoCardHeader.module.scss
@@ -1,16 +1,0 @@
-.cardHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 12px 16px 16px 0;
-
-  @media screen and (min-width: 768px) {
-    padding: 16px 20px 20px 0;
-  }
-}
-
-// .heartIcon {
-//   height: 40px;
-//   width: auto;
-// }
