+++ front/src/features/idea-memos-uuid/components/index.ts
@@ -1,21 +1,17 @@
 import ButtonSection from "@/features/idea-memos-uuid/components/ButtonSection/ButtonSection";
 import CancelButton from "@/features/idea-memos-uuid/components/CancelButton/CancelButton";
 import CardBodyOnViewMode from "@/features/idea-memos-uuid/components/CardBodyOnViewMode/CardBodyOnViewMode";
-import CardFooter from "@/features/idea-memos-uuid/components/CardFooter/CardFooter";
 import DeleteButton from "@/features/idea-memos-uuid/components/DeleteButton/DeleteButton";
 import EditButton from "@/features/idea-memos-uuid/components/EditButton/EditButton";
 import EditForm from "@/features/idea-memos-uuid/components/EditForm/EditForm";
 import IdeaMemoCardContent from "@/features/idea-memos-uuid/components/IdeaMemoCardContent/IdeaMemoCardContent";
-import IdeaMemoCardHeader from "@/features/idea-memos-uuid/components/IdeaMemoCardHeader/IdeaMemoCardHeader";
 
 export {
   ButtonSection,
   CancelButton,
   CardBodyOnViewMode,
-  CardFooter,
   DeleteButton,
   EditButton,
   EditForm,
   IdeaMemoCardContent,
-  IdeaMemoCardHeader,
 };
